add new like list ui and refactor with modular css
