+++ src/styles/base.css
@@ -1,0 +1,58 @@
+/* normalize */
+
+html, body {
+  height: 100%;
+  font-size: var(--base-font-size);
+  line-height: 1.5;
+  font-family: Inter;
+  color: var(--primary-100);
+}
+
+main {
+  min-height: calc(100vh - 56px - 99px);
+}
+
+img {
+  max-width: 100%;
+  height: auto;
+  display: block;
+}
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
+a {
+  text-decoration: none;
+  color: inherit;
+}
+
+/* Typography */
+
+h1, h2, h3, h4, h5, h6 {
+  line-height: 1.1;
+  margin-bottom: 1rem;
+  color: var(--accent-500);
+  font-style: italic;
+  font-weight: bold;
+}
+
+h1 {
+  font-size: var(--h1-font-size);
+}
+
+h2 {
+  font-size: var(--h2-font-size);
+}
+
+h3 {
+  font-size: var(--h3-font-size);
+}
+
+h4 {
+  font-size: var(--h4-font-size);
+}
+
+
