+++ src/styles/components.css
@@ -1,0 +1,119 @@
+button {
+  border-radius: 8px;
+  border: 1px solid transparent;
+  padding: 1em 2em;
+  font-size: 1em;
+  font-weight: 500;
+  font-family: inherit;
+  background-color: var(--primary-color);
+  color: #fff;
+  cursor: pointer;
+  transition: border-color 0.25s;
+}
+button:hover {
+  filter: brightness(0.8);
+}
+button:focus,
+button:focus-visible {
+  outline: 4px auto -webkit-focus-ring-color;
+}
+
+
+.button-primary {
+  width: clamp(120px, 100%, 600px);
+  height: clamp(40px, 100%, 60px);
+  background-color: var(--secondary-500);
+  border: 2px solid var(--secondary-500);
+  color: #fff;
+  border-radius: 32px;
+  font-weight: bold;
+  font-style: italic;
+}
+  
+.button-outline {
+  width: clamp(120px, 100%, 600px);
+  height: clamp(40px, 100%, 60px);
+  background-color: transparent;
+  border: 2px solid white;
+  color: white;
+  cursor: pointer;
+  border-radius: 32px;
+  font-weight: bold;
+  font-style: italic;
+}
+
+.button-info {
+  background-color: white;
+  border: 1px solid var(--secondary-500);
+  color: var(--secondary-500);
+  border-radius: 100%;
+  height: 32px;
+  width: 32px;
+  padding: 0;
+  font-weight: bold;
+}
+
+.button-icon {
+  border: 2px solid var(--primary-100);
+  color: var(--primary-100);
+  border-radius: 100%;
+  height: 48px;
+  width: 48px;
+  padding: 0;
+  font-weight: bold;
+}
+
+.button-circle {
+  border: 2px solid white;
+  color: white;
+  border-radius: 100%;
+  height: 64px;
+  width: 64px;
+  padding: 0;
+  font-weight: bold;
+}
+
+/* inputs */
+
+input[type="text"],
+input[type="email"],
+input[type="password"],
+textarea {
+  border: 2px solid var(--primary-100);
+  color: white;
+  background-color: transparent;
+  border-radius: 48px;
+  height: 48px;
+  padding: 0 var(--gap-lg);
+}
+
+input:focus,
+textarea:focus {
+  outline: 4px auto -webkit-focus-ring-color;
+}
+
+.filter-bar {
+    display: grid;
+    grid-template-columns: 1fr 48px;
+    gap: var(--gap-sm);
+    margin-bottom: var(--gap-md);
+}
+
+/* Header */
+
+.logo {
+  font-size: 2rem;
+  margin: 1rem;
+  color: white;
+}
+
+/* footer */
+
+footer {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 1rem 2rem;
+  background-color: var(--primary-700);
+  color: white;
+}
