+++ src/styles/layouts.css
@@ -1,0 +1,61 @@
+/* grids */
+
+.grid {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: var(--gap-md);
+}
+
+.grid-cols-2 {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+}
+
+.grid-cols-3 {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+}
+
+/* containers */
+
+.container-sm {
+    max-width: 440px;
+}
+
+.container-md {
+    max-width: 768px;
+}
+
+.container-lg {
+    max-width: 1024px;
+}
+
+.container-xl {
+    max-width: 1520px;
+}
+
+.container-full {
+    max-width: 100%;
+}
+
+
+/* header */
+header {
+  padding: 1rem 2rem;
+  background-color: var(--primary-700);
+}
+
+nav {
+  display: flex;
+  gap: 1rem;
+}
+
+/* body */
+main {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 2rem;
+  background-color: var(--primary-500);
+}
