+++ src/styles/utils.css
@@ -1,0 +1,219 @@
+/* colors */
+
+.bg-primary {
+    background-color: var(--primary-color);
+}
+
+.bg-secondary {
+    background-color: var(--secondary-color);
+}
+
+.bg-white {
+    background-color: #fff;
+}
+
+.text-primary {
+    color: var(--text-primary);
+}
+
+.text-secondary {
+    color: var(--text-secondary);
+}
+
+.text-white {
+    color: #fff;
+}
+
+/* margin */
+
+.m-0 {
+    margin: 0;
+}
+
+.m-1 {
+    margin: var(--gap-sm);
+}
+
+.m-2 {
+    margin: var(--gap-md);
+}
+
+.m-3 {
+    margin: var(--gap-lg);
+}
+
+/* padding */
+
+.p-0 {
+    padding: 0;
+}
+
+.p-1 {
+    padding: var(--gap-sm);
+}
+
+.p-2 {
+    padding: var(--gap-md);
+}
+
+.p-3 {
+    padding: var(--gap-lg);
+}
+
+.px-1 {
+    padding-left: var(--gap-sm);
+    padding-right: var(--gap-sm);
+}
+
+.px-2 {
+    padding-left: var(--gap-md);
+    padding-right: var(--gap-md);
+}
+
+.px-3 {
+    padding-left: var(--gap-lg);
+    padding-right: var(--gap-lg);
+}
+
+.py-1 {
+    padding-top: var(--gap-sm);
+    padding-bottom: var(--gap-sm);
+}
+
+.py-2 {
+    padding-top: var(--gap-md);
+    padding-bottom: var(--gap-md);
+}
+
+.py-3 {
+    padding-top: var(--gap-lg);
+    padding-bottom: var(--gap-lg);
+}
+
+/* sizing */
+
+.full {
+    width: 100%;
+}
+
+.half {
+    width: 50%;
+}
+
+.third {
+    width: 33.33%;
+}
+
+
+/* flex */
+
+.flex-row {
+  display: flex;
+  flex-direction: row;
+}
+
+.flex-col {
+  display: flex;
+  flex-direction: column;
+}
+
+.justify-start {
+  justify-content: start;
+}
+.justify-center {
+  justify-content: center;
+}
+.justify-between {
+  justify-content: space-between;
+}
+.justify-end {
+  justify-content: end;
+}
+
+.justify-around {
+  justify-content: space-around;
+}
+
+.items-start {
+  align-items: start;
+}
+.items-center {
+  align-items: center;
+}
+.items-end {
+  align-items: end;
+}
+
+/* gap */
+
+.gap-1 {
+  gap: var(--gap-sm);
+}
+
+.gap-2 {
+  gap: var(--gap-md);
+}
+
+.gap-3 {
+  gap: var(--gap-lg);
+}
+
+.gap-4 {
+  gap: var(--gap-xl);
+}
+
+/* display */
+
+.block {
+  display: block;
+}
+
+.inline-block {
+  display: inline-block;
+}
+
+.hidden {
+  display: none;
+}
+
+/* positioning */
+
+.relative {
+  position: relative;
+}
+
+.absolute {
+  position: absolute;
+}
+
+.fixed {
+  position: fixed;
+}
+
+.sticky {
+  position: sticky;
+}
+
+.pos-bottom-right {
+  position: absolute;
+  bottom: 0;
+  right: 0;
+}
+
+.pos-bottom-left {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+}
+
+.pos-top-right {
+  position: absolute;
+  top: 0;
+  right: 0;
+}
+
+.pos-top-left {
+  position: absolute;
+  top: 0;
+  left: 0;
+}
